[BugFix] World scaling & frustum culling.
Now the rendering is pretty fast, but still have problems with performance, the amount of vertices make the rendering performance very poor, I will use 2D quads for far trees.

Other issue was with world scaling, the identity heightmap (empty heightmap mesh) was storing 1m*(w, h) which was wrong, now is the heightmap is 1m for w and h.

1- Fixed frustum culling.
2- Best performance because frustum culling is working.
3- World scale is fixed.